Update LoginActivity : Seperated LoginView
